requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.0.1
bleach==3.2.1
blinker==1.4
boto3==1.16.23
botocore==1.19.23
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cmake==3.18.4.post1
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dlib @ file:///D:/bld/dlib_1602513956312/work
entrypoints==0.3
enum-compat==0.0.3
gitdb==4.0.5
GitPython==3.1.11
idna==2.10
imageio==2.9.0
imutils==0.5.3
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.5
numpy @ file:///C:/ci/numpy_and_numpy_base_1603466732592/work
olefile==0.46
opencv-contrib-python==4.4.0.46
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pathtools==0.1.2
pickleshare==0.7.5
pilgram==1.2.0
Pillow @ file:///C:/ci/pillow_1603823068645/work
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
pyarrow==2.0.0
pycparser==2.20
pydeck==0.5.0
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
pyzmq==20.0.0
requests==2.25.0
s3transfer==0.3.3
scikit-image==0.17.2
scipy==1.5.4
Send2Trash==1.5.0
six @ file:///C:/ci/six_1605187374963/work
smmap==3.0.4
streamlit==0.71.0
terminado==0.9.1
testpath==0.4.4
tifffile==2020.11.18
toml==0.10.2
toolz==0.11.1
torch==1.7.0
torchvision==0.8.1
tornado==6.1
traitlets==5.0.5
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
tzlocal==2.1
urllib3==1.26.2
validators==0.18.1
watchdog==0.10.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wincertstore==0.2
